introduce second level of abstraction
