+++ server/src/main/java/org/hl7/davinci/endpoint/cdshooks/services/crd/CdsService.java
@@ -1,31 +1,19 @@
 package org.hl7.davinci.endpoint.cdshooks.services.crd;
 
-import java.net.MalformedURLException;
-import java.net.URI;
-import java.net.URL;
-import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
-import java.util.List;
-import java.util.ArrayList;
-import java.util.Date;
-import javax.validation.Valid;
-
 import org.apache.commons.lang.StringUtils;
 import org.cdshooks.*;
 import org.hl7.davinci.FhirComponentsT;
 import org.hl7.davinci.PrefetchTemplateElement;
 import org.hl7.davinci.RequestIncompleteException;
-import org.hl7.davinci.endpoint.config.YamlConfig;
 import org.hl7.davinci.endpoint.components.CardBuilder;
 import org.hl7.davinci.endpoint.components.CardBuilder.CqlResultsForCard;
 import org.hl7.davinci.endpoint.components.PrefetchHydrator;
 import org.hl7.davinci.endpoint.database.RequestLog;
 import org.hl7.davinci.endpoint.database.RequestService;
 import org.hl7.davinci.endpoint.files.FileStore;
-import org.hl7.davinci.endpoint.rules.CoverageRequirementRuleCriteria;
 import org.hl7.davinci.endpoint.rules.CoverageRequirementRuleResult;
-import org.hl7.davinci.r4.crdhook.orderselect.OrderSelectRequest;
 import org.hl7.davinci.r4.crdhook.DiscoveryExtension;
+import org.hl7.davinci.r4.crdhook.orderselect.OrderSelectRequest;
 import org.opencds.cqf.cql.engine.execution.Context;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -33,41 +21,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.RequestBody;
 
+import javax.validation.Valid;
+import java.net.URL;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
+
 @Component
-public abstract class CdsService<requestTypeT extends CdsRequest<?, ?>> {
+public abstract class CdsService<requestTypeT extends CdsRequest<?, ?>> extends CdsAbstract<requestTypeT> {
   static final Logger logger = LoggerFactory.getLogger(CdsService.class);
-
-  /**
-   * The {id} portion of the URL to this service which is available at
-   * {baseUrl}/cds-services/{id}. REQUIRED
-   */
-  public String id;
-
-  /**
-   * The hook this service should be invoked on. REQUIRED
-   */
-  public Hook hook;
-
-  /**
-   * The human-friendly name of this service. RECOMMENDED
-   */
-  public String title;
-
-  /**
-   * The description of this service. REQUIRED
-   */
-  public String description;
-
-  /**
-   * An object containing key/value pairs of FHIR queries that this service is
-   * requesting that the EHR prefetch and provide on each service call. The key is
-   * a string that describes the type of data being requested and the value is a
-   * string representing the FHIR query. OPTIONAL
-   */
-  public Prefetch prefetch;
-
-  @Autowired
-  private YamlConfig myConfig;
 
   @Autowired
   RequestService requestService;
@@ -75,54 +37,10 @@ public abstract class CdsService<requestTypeT extends CdsRequest<?, ?>> {
   @Autowired
   FileStore fileStore;
 
-  private final List<PrefetchTemplateElement> prefetchElements;
-
-  protected FhirComponentsT fhirComponents;
-
-  private final DiscoveryExtension extension;
-
-  /**
-   * Create a new cdsservice.
-   *
-   * @param id               Will be used in the url, should be unique.
-   * @param hook             Which hook can call this.
-   * @param title            Human title.
-   * @param description      Human description.
-   * @param prefetchElements List of prefetch elements, will be in prefetch
-   *                         template.
-   * @param fhirComponents   Fhir components to use
-   * @param extension        Custom CDS Hooks extensions.
-   */
   public CdsService(String id, Hook hook, String title, String description,
       List<PrefetchTemplateElement> prefetchElements, FhirComponentsT fhirComponents,
       DiscoveryExtension extension) {
-
-    if (id == null) {
-      throw new NullPointerException("CDSService id cannot be null");
-    }
-    if (hook == null) {
-      throw new NullPointerException("CDSService hook cannot be null");
-    }
-    if (description == null) {
-      throw new NullPointerException("CDSService description cannot be null");
-    }
-    this.id = id;
-    this.hook = hook;
-    this.title = title;
-    this.description = description;
-    this.prefetchElements = prefetchElements;
-    prefetch = new Prefetch();
-    for (PrefetchTemplateElement prefetchElement : prefetchElements) {
-      this.prefetch.put(prefetchElement.getKey(), prefetchElement.getQuery());
-    }
-    this.fhirComponents = fhirComponents;
-    this.extension = extension;
-  }
-
-  public DiscoveryExtension getExtension() { return extension; }
-
-  public List<PrefetchTemplateElement> getPrefetchElements() {
-    return prefetchElements;
+    super(id, hook, title, description, prefetchElements, fhirComponents, extension);
   }
 
   /**
@@ -239,7 +157,7 @@ public abstract class CdsService<requestTypeT extends CdsRequest<?, ?>> {
       CardBuilder.errorCardIfNonePresent(response);
     }
 
-    // Ading card to requestLog
+    // Adding card to requestLog
     requestLog.setCardListFromCards(response.getCards());
     requestService.edit(requestLog);
 
@@ -304,68 +222,7 @@ public abstract class CdsService<requestTypeT extends CdsRequest<?, ?>> {
     }
   }
 
-  private Link smartLinkBuilder(String patientId, String fhirBase, URL applicationBaseUrl, String questionnaireUri,
-      String reqResourceId, CoverageRequirementRuleCriteria criteria, boolean priorAuthRequired, String label) {
-    URI configLaunchUri = myConfig.getLaunchUrl();
-    questionnaireUri = applicationBaseUrl + "/fhir/r4/" + questionnaireUri;
 
-    String launchUrl;
-    if (myConfig.getLaunchUrl().isAbsolute()) {
-      launchUrl = myConfig.getLaunchUrl().toString();
-    } else {
-      try {
-        launchUrl = new URL(applicationBaseUrl.getProtocol(), applicationBaseUrl.getHost(),
-            applicationBaseUrl.getPort(), applicationBaseUrl.getFile() + configLaunchUri.toString(), null).toString();
-      } catch (MalformedURLException e) {
-        String msg = "Error creating smart launch URL";
-        logger.error(msg);
-        throw new RuntimeException(msg);
-      }
-    }
-
-    if (fhirBase != null && fhirBase.endsWith("/")) {
-      fhirBase = fhirBase.substring(0, fhirBase.length() - 1);
-    }
-    if (patientId != null && patientId.startsWith("Patient/")) {
-      patientId = patientId.substring(8);
-    }
-
-    // PARAMS:
-    // template is the uri of the questionnaire
-    // request is the ID of the device request or medrec (not the full URI like the
-    // IG says, since it should be taken from fhirBase
-
-    String filepath = "../../getfile/" + criteria.getQueryString();
-
-    String appContext = "template=" + questionnaireUri + "&request=" + reqResourceId;
-    appContext = appContext + "&fhirpath=" + applicationBaseUrl + "/fhir/";
-
-    appContext = appContext + "&priorauth=" + (priorAuthRequired ? "true" : "false");
-    appContext = appContext + "&filepath=" + applicationBaseUrl + "/";
-    if (myConfig.getUrlEncodeAppContext()) {
-      logger.info("CdsService::smartLinkBuilder: URL encoding appcontext");
-      appContext = URLEncoder.encode(appContext, StandardCharsets.UTF_8).toString();
-    }
-
-    logger.info("smarLinkBuilder: appContext: " + appContext);
-
-    if (myConfig.isAppendParamsToSmartLaunchUrl()) {
-      launchUrl = launchUrl + "?iss=" + fhirBase + "&patientId=" + patientId + "&template=" + questionnaireUri
-          + "&request=" + reqResourceId;
-    } else {
-      // TODO: The iss should be set by the EHR?
-      launchUrl = launchUrl;
-    }
-
-    Link link = new Link();
-    link.setType("smart");
-    link.setLabel(label);
-    link.setUrl(launchUrl);
-
-    link.setAppContext(appContext);
-
-    return link;
-  }
 
   // Implement these in child class
   public abstract List<CoverageRequirementRuleResult> createCqlExecutionContexts(requestTypeT request,
